Simplify usage of os/exec
For now, keep the finishRunning() wrapper but use a straight cmd.Run()
call instead of the convoluted goroutine trying to catch signals.

It turns out that Unix process group handling is enough to interrupt
pending processes when stopping the run with something like a Ctrl+C
which should be enough.

Tested:
- Full e2e run with hack/e2e-test.sh, two tests failed but looks like
  they've been failing before this change.
- Started a hack/e2e.go -v -build and interrupted it with Ctrl+C,
  confirmed that build-release.sh was killed in the process.