Update conformance test.
Fix formatting and quote variable.
Fix comments from a previous review.
Move instructions to the top.
Change from using tag to using branch.
Improve documentation of conformance test.
Use stable random seed.
Fix bug where there was a double quote as a prefix and suffix to the
skip regexp.
Remove .* not needed in regex.
Skip a few more tests that are not reliable for me in the
Conformance test environment.