Remove cgroup list dependent e2e test from conformance test suite
As discussed @gmarek the given test does not belong into the conformance test
suite because it makes a lot of static assumptions about the cgroup setup of the
nodes which cannot be fulfilled by all cluster providers. Depending on the
installation the kubelet is not allowed to move around process
into specific containers.

Fixes https://github.com/mesosphere/kubernetes-mesos/issues/439.