Fix Vagrant issue with Salt
Package manager "dnf" does not work correctly with Salt
(cf https://github.com/saltstack/salt/issues/31001)

It causes Salt to consider that some packages (python, git, curl, etc.) are not
installed, which breaks the Vagrant Kubernetes setup.

Updating dnf and dnf-plugins-core to their latest version solves the issue.

Additionally, I've added the "fastestmirror" to dnf, which is useful if a
RPM mirror is broken or very slow. (In my case, dnf used a broken mirror which
froze the Kubernetes setup).