Call 'gcloud compute' only if necessary
The 'gcloud compute' is called everytime even if it is not needed.
When runnin node e2e tests on RHEL, the test-e2e-node.sh script is run
the way in which it takes execution path without gcloud invocation.

With the current code, gcloud is called everytime.
Thus introducing additional runtime dependency.